Replace unzipper with yauzl
The library [unzipper][1] was originally used to unzip a stream without
the need to write the zipfile to disk. While this is nice the zip format
itself is not really designed to be streamed.

It still seems to work, but the `unzipper` package also a lot of
dependencies as well.

We are now moving to a different library [yauzl][2]. This only works
with complete files, so we have to download the zipfile before we can
extract it. This library only has 2 dependencies with no further
dependencies.

[1]: https://www.npmjs.com/package/unzipper
[2]: https://www.npmjs.com/package/yauzl